More work for Issue #6, which is to add save/modify.
A save(DirectoryServiceEntry entry) method has been added to DirectoryService, and while working through this realized that some more plubming needed to be in the DirectoryServiceEntry class, like the original baseDN which was used to find the object, an errors map to keep track of errors, and a method for cleaning things up after a successful save.

Test cases have been added to support this new feature.